Create AdminPeriodController
