+++ app/Resources/views/admin/index.html.twig
@@ -68,7 +68,7 @@
 
             {% if is_granted("ROLE_SHIFT_MANAGER") %}
                 <h6><i class="material-icons">date_range</i>&nbsp;Semaine type</h6>
-                <a href="{{ path("period_admin") }}" class="waves-effect waves-light btn white black-text">
+                <a href="{{ path("admin_period_index") }}" class="waves-effect waves-light btn white black-text">
                     <i class="material-icons left orange-text">date_range</i>Gérer la semaine type
                 </a>
                 <a href="{{ path("admin_periodpositionfreelog_list") }}" class="waves-effect waves-light btn orange">
