+++ app/Resources/views/admin/period/index.html.twig
@@ -101,7 +101,7 @@
                     </div>
                     <div class="collapsible-body">
                         <div class="row">
-                            <a href="{{ path("period_new") }}"
+                            <a href="{{ path("admin_period_new") }}"
                                class="btn col m4 waves-effect waves-light tooltipped"
                                data-position="bottom"
                                data-tooltip="Créer un groupe de créneaux type à une heure et un jour donné">
@@ -111,7 +111,7 @@
 
                         {% if is_granted("ROLE_ADMIN") %}
                             <div class="row">
-                                <a href="{{ path("period_copy") }}"
+                                <a href="{{ path("admin_period_copy") }}"
                                    class="btn col m4 waves-effect waves-light tooltipped"
                                    data-position="bottom"
                                    data-tooltip="Dupliquer des créneaux types d'un jour vers un autre">
@@ -119,7 +119,7 @@
                                 </a>
                             </div>
                             <div class="row">
-                                <a href="{{ path("shifts_generation") }}"
+                                <a href="{{ path("admin_shifts_generation") }}"
                                    class="btn col m4 waves-effect waves-light tooltipped"
                                    data-position="bottom"
                                    data-tooltip="Générer automatiquement des créneaux à partir de la semaine type">
