+++ app/Resources/views/default/index.html.twig
@@ -127,7 +127,7 @@
                     <i class="material-icons left">schedule</i>Planning
                 </a>
                 <br />
-                <a href="{{ path("period") }}" class="waves-effect waves-light btn teal">
+                <a href="{{ path("period_index") }}" class="waves-effect waves-light btn teal">
                     <i class="material-icons left">date_range</i>Semaine type
                 </a>
                 <br />
