+++ app/Resources/views/period/_partial/period_card.html.twig
@@ -29,7 +29,7 @@
                 {{ _self.period_title(period, show_day_of_week) }}
             </div>
         {% else %}
-            <a href="{{ path("period_edit",{'id': period.id}) }}" class="black-text">
+            <a href="{{ path("admin_period_edit",{'id': period.id}) }}" class="black-text">
                 <div class="black-text editable-box">
                     {{ _self.period_title(period, show_day_of_week) }}
                 </div>
