+++ app/Resources/views/user/_partial/period_position_card.html.twig
@@ -35,7 +35,7 @@
             </span>
         </div>
         <div class="card-action">
-            <a href="{{ path("period_edit", { 'id': period_position.period.id }) }}" class="waves-effect waves-light btn white black-text" target="_blank" title="Voir le créneau">
+            <a href="{{ path("admin_period_edit", { 'id': period_position.period.id }) }}" class="waves-effect waves-light btn white black-text" target="_blank" title="Voir le créneau">
                 <i class="material-icons left orange-text">date_range</i>Voir le créneau
             </a>
         </div>
